+++ components/notification-mgt/io.entgra.device.mgt.core.notification.mgt.common/src/main/java/io/entgra/device/mgt/core/notification/mgt/common/service/NotificationConfigService.java
@@ -12,16 +12,10 @@ public interface NotificationConfigService {
      * @throws NotificationConfigurationServiceException Throws when error occurred while retrieving notifications.
      */
         void addNotificationConfigContext(NotificationConfigurationList configurations) throws NotificationConfigurationServiceException;
-
-
         void deleteNotificationConfigContext(String configID) throws NotificationConfigurationServiceException;
-
         void updateNotificationConfigContext(NotificationConfig updatedConfig) throws NotificationConfigurationServiceException;
-
         void deleteNotificationConfigurations() throws NotificationConfigurationServiceException;
-
         NotificationConfigurationList getNotificationConfigurations() throws NotificationConfigurationServiceException;
-
         NotificationConfig getNotificationConfigByID(String configID) throws NotificationConfigurationServiceException;
